fix:一秒以下の動画の投稿が終わらない
フレームレートの計算時に
duration を nano sec から sec に変換するが、
その際に float 化しそこねていたのが原因。

+++ library/src/com/lisb/android/mediashrink/VideoShrink.java
@@ -142,8 +142,8 @@ public class VideoShrink {
 		format.setInteger(MediaFormat.KEY_BIT_RATE, bitRate);
 
 		if (frameCount > 0) {
-			final float durationSec = origin.getLong(MediaFormat.KEY_DURATION)
-					/ (1000 * 1000);
+			final float durationSec = (float) origin
+					.getLong(MediaFormat.KEY_DURATION) / (1000 * 1000);
 			float frameRate = frameCount / durationSec;
 
 			Log.d(LOG_TAG, "video frame-count:" + frameCount + ", frame-rate:"
